requirements.txt
curses
tabulate
colorama
mysql-connector-python